requirements.txt
tensorflow
flask
Pillow
cassandra-driver